+++ export_blocks_and_transactions.py
@@ -26,7 +26,7 @@ import argparse
 from ethereumetl.jobs.export_blocks_job import ExportBlocksJob
 from ethereumetl.jobs.export_blocks_job_item_exporter import export_blocks_job_item_exporter
 from ethereumetl.thread_local_proxy import ThreadLocalProxy
-from ethereumetl.web3_utils import get_batch_provider_from_uri
+from ethereumetl.providers.auto import get_batch_provider_from_uri
 
 parser = argparse.ArgumentParser(description='Export blocks and transactions.')
 parser.add_argument('-s', '--start-block', default=0, type=int, help='Start block')
