+++ requirements.txt
@@ -1,3 +1,3 @@
-web3~=4.3.0
+web3~=4.4.1
 pytest~=3.2.0
 python-dateutil~=2.7.0
